+++ lib/src/theme/styles/styles.dart
@@ -7,6 +7,8 @@ export 'src/sbb_chip_style.dart';
 export 'src/sbb_control_style.dart';
 export 'src/sbb_control_styles.dart';
 export 'src/sbb_headerbox_style.dart';
+export 'src/sbb_group_style.dart';
+export 'src/sbb_headerbox_style.dart';
 export 'src/sbb_list_item_style.dart';
 export 'src/sbb_pagination_style.dart';
 export 'src/sbb_picker_style.dart';