+++ test/sbb_toast_test.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 import 'package:sbb_design_system_mobile/src/toast/default_toast_body.dart';
 
 import 'test_app.dart';
@@ -11,15 +12,19 @@ void main() {
   final Stream<bool> stream1 = (StreamController<bool>()..add(true)).stream;
   const Duration duration = Duration(milliseconds: 0);
   testWidgets('toast basic test', (WidgetTester tester) async {
-    final widget = Column(
-      children: [
-        DefaultToastBody(
-          title: title,
-          duration: duration,
-          stream: stream1,
-        ),
-      ],
-    );
+    final widget = Builder(builder: (context) {
+      final sbbToast = SBBToast.of(context);
+      return Column(
+        children: [
+          DefaultToastBody(
+            title: title,
+            duration: duration,
+            stream: stream1,
+            toast: sbbToast,
+          ),
+        ],
+      );
+    });
 
     await TestSpecs.run(
       TestSpecs.themedSpecs,