+++ CHANGELOG.md
@@ -11,7 +11,18 @@ It is expected that you keep this format strictly, since we depend on it in our 
 
 ### Changed
 
+- BREAKING: `SBBGroup` does not have a shadow anymore
 - updated icon lib to version 1.7.1
+- Simpler implementation of the `SBBTabBar`.
+- `SBBTabBar` now respects the `SafeArea` for the `systemNavigationBar`.
+
+### Added
+
+- `SBBGroupStyle` is used to style `SBBGroup` in a theme extension like manner
+
+### Removed
+
+- removed `groupBackgroundColor` from `SBBControlStyles`
 
 ## [3.2.0] - 2025-06-23